+++ src/format/mod.rs
@@ -47,9 +47,8 @@ use std::error::Error;
 
 #[cfg(any(feature = "alloc", feature = "std", test))]
 use crate::naive::{NaiveDate, NaiveTime};
-use crate::offset::FixedOffset;
 #[cfg(any(feature = "alloc", feature = "std", test))]
-use crate::offset::Offset;
+use crate::offset::{FixedOffset, Offset};
 #[cfg(any(feature = "alloc", feature = "std", test))]
 use crate::{Datelike, Timelike};
 use crate::{Month, ParseMonthError, ParseWeekdayError, Weekday};
@@ -421,7 +420,7 @@ impl Error for ParseError {
 }
 
 // to be used in this module and submodules
-const OUT_OF_RANGE: ParseError = ParseError(ParseErrorKind::OutOfRange);
+pub(crate) const OUT_OF_RANGE: ParseError = ParseError(ParseErrorKind::OutOfRange);
 const IMPOSSIBLE: ParseError = ParseError(ParseErrorKind::Impossible);
 const NOT_ENOUGH: ParseError = ParseError(ParseErrorKind::NotEnough);
 const INVALID: ParseError = ParseError(ParseErrorKind::Invalid);
@@ -839,7 +838,7 @@ mod parsed;
 
 // due to the size of parsing routines, they are in separate modules.
 mod parse;
-mod scan;
+pub(crate) mod scan;
 
 pub mod strftime;
 
@@ -989,15 +988,6 @@ impl FromStr for Weekday {
         } else {
             Err(ParseWeekdayError { _dummy: () })
         }
-    }
-}
-
-/// Parsing a `str` into a `FixedOffset` uses the format [`%z`](./format/strftime/index.html).
-impl FromStr for FixedOffset {
-    type Err = ParseError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let (_, offset) = scan::timezone_offset(s, scan::colon_or_space)?;
-        Self::east_opt(offset).ok_or(OUT_OF_RANGE)
     }
 }
